refactored half include player and weapon system

+++ Assets/_Project/Scripts/ScriptableObjects/Configs/EnemyConfig.cs
@@ -8,7 +8,7 @@ namespace Assets._Project.Scripts.ScriptableObjects.Configs
         [field: SerializeField] public Enemy.Enemy Prefab;
         [field: SerializeField] public int Health;
         [field: SerializeField] public int Damage;
-        [field: SerializeField] public int Speed;
+        [field: SerializeField] public float Speed;
         [field: SerializeField] public float AttackRadius;
         [field: SerializeField] public float VisibilityRadius;
         [field: SerializeField] public int ReturnInitialAttackPosition;
